refactor(entry): move internal/entry to system/entry
The entry config decoder now depends on the api/env contract and owns
env-resolution policy, so it belongs in system alongside the payload,
env, and registry implementations rather than behind the internal
barrier. DecodeEntryConfig becomes public subsystem API.
