+++ service/aws/sqs/manager.go
@@ -15,8 +15,8 @@ import (
 	"github.com/wippyai/runtime/api/resource"
 	sqsapi "github.com/wippyai/runtime/api/service/aws/sqs"
 	"github.com/wippyai/runtime/api/supervisor"
-	entryutil "github.com/wippyai/runtime/internal/entry"
 	queuesvc "github.com/wippyai/runtime/service/queue"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	"go.uber.org/zap"
 )
