+++ service/fs/embed/manager.go
@@ -12,7 +12,7 @@ import (
 	"github.com/wippyai/runtime/api/payload"
 	"github.com/wippyai/runtime/api/registry"
 	embedapi "github.com/wippyai/runtime/api/service/fs/embed"
-	entryutil "github.com/wippyai/runtime/internal/entry"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	systemfs "github.com/wippyai/runtime/system/fs"
 	"go.uber.org/zap"
 )
