+++ service/pg/manager.go
@@ -24,7 +24,7 @@ import (
 	pgapi "github.com/wippyai/runtime/api/service/pg"
 	"github.com/wippyai/runtime/api/supervisor"
 	"github.com/wippyai/runtime/api/topology"
-	entryutil "github.com/wippyai/runtime/internal/entry"
+	entryutil "github.com/wippyai/runtime/system/entry"
 
 	"go.opentelemetry.io/otel"
 	"go.uber.org/zap"
