+++ service/queue/amqp/manager.go
@@ -12,8 +12,8 @@ import (
 	"github.com/wippyai/runtime/api/registry"
 	amqpapi "github.com/wippyai/runtime/api/service/queue/amqp"
 	"github.com/wippyai/runtime/api/supervisor"
-	entryutil "github.com/wippyai/runtime/internal/entry"
 	queuesvc "github.com/wippyai/runtime/service/queue"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	"go.uber.org/zap"
 )
