+++ service/sql/manager.go
@@ -16,7 +16,7 @@ import (
 	"github.com/wippyai/runtime/api/resource"
 	config "github.com/wippyai/runtime/api/service/sql"
 	"github.com/wippyai/runtime/api/supervisor"
-	entryutil "github.com/wippyai/runtime/internal/entry"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	"go.uber.org/zap"
 )
