+++ service/store/memory/manager.go
@@ -12,8 +12,8 @@ import (
 	"github.com/wippyai/runtime/api/resource"
 	memstore "github.com/wippyai/runtime/api/service/store/memory"
 	"github.com/wippyai/runtime/api/supervisor"
-	entryutil "github.com/wippyai/runtime/internal/entry"
 	storesvc "github.com/wippyai/runtime/service/store"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	"go.uber.org/zap"
 )
