+++ service/store/sql/manager.go
@@ -12,8 +12,8 @@ import (
 	"github.com/wippyai/runtime/api/resource"
 	sqlstore "github.com/wippyai/runtime/api/service/store/sql"
 	"github.com/wippyai/runtime/api/supervisor"
-	entryutil "github.com/wippyai/runtime/internal/entry"
 	storesvc "github.com/wippyai/runtime/service/store"
+	entryutil "github.com/wippyai/runtime/system/entry"
 	"go.uber.org/zap"
 )
