+++ system/entry/decode.go
