+++ system/entry/decode_resolve_test.go
