+++ system/entry/decode_test.go
