+++ system/entry/errors.go
