+++ system/entry/resolve.go
